refactor: move `app/livechat` and `app/livechat-enterprise` modules (#41920)
