--- apps/meteor/app/livechat-enterprise/client/index.ts
@@ -1,9 +1,0 @@
-import { hasLicense } from '../../license/client';
-
-void hasLicense('livechat-enterprise').then((enabled) => {
-	if (!enabled) {
-		return;
-	}
-
-	return Promise.all([import('./views/livechatSideNavItems'), import('./views/business-hours/Multiple')]);
-});
