+++ apps/meteor/client/lib/omnichannel/businessHours/MultipleBusinessHoursBehavior.ts
@@ -1,7 +1,7 @@
 import type { ILivechatBusinessHour } from '@rocket.chat/core-typings';
 import { LivechatBusinessHourTypes } from '@rocket.chat/core-typings';
 
-import type { IBusinessHourBehavior } from '../../../../livechat/client/views/app/business-hours/IBusinessHourBehavior';
+import type { IBusinessHourBehavior } from './IBusinessHourBehavior';
 
 export class MultipleBusinessHoursBehavior implements IBusinessHourBehavior {
 	getView(): string {
