+++ apps/meteor/client/views/omnichannel/businessHours/useIsSingleBusinessHours.ts
@@ -1,4 +1,4 @@
-import { businessHourManager } from '../../../../app/livechat/client/views/app/business-hours/BusinessHours';
+import { businessHourManager } from '../../../lib/omnichannel/businessHours/businessHourManager';
 
 // businessHourManager holds a single behavior instance with no reactive store — the previous
 // useReactiveValue wrapper never actually re-ran, so this is a plain read.
