+++ apps/meteor/client/views/root/hooks/useLivechatEnterprise.ts
@@ -1,24 +1,23 @@
 import { useSetting } from '@rocket.chat/ui-contexts';
 import { useEffect } from 'react';
 
-import { businessHourManager } from '../../../../app/livechat/client/views/app/business-hours/BusinessHours';
-import type { IBusinessHourBehavior } from '../../../../app/livechat/client/views/app/business-hours/IBusinessHourBehavior';
-import { SingleBusinessHourBehavior } from '../../../../app/livechat/client/views/app/business-hours/Single';
-import { MultipleBusinessHoursBehavior } from '../../../../app/livechat-enterprise/client/views/business-hours/Multiple';
 import { useHasLicenseModule } from '../../../hooks/useHasLicenseModule';
+import { MultipleBusinessHoursBehavior } from '../../../lib/omnichannel/businessHours/MultipleBusinessHoursBehavior';
+import { SingleBusinessHourBehavior } from '../../../lib/omnichannel/businessHours/SingleBusinessHourBehavior';
+import { businessHourManager } from '../../../lib/omnichannel/businessHours/businessHourManager';
 
-const businessHours: Record<string, IBusinessHourBehavior> = {
+const businessHours = {
 	multiple: new MultipleBusinessHoursBehavior(),
 	single: new SingleBusinessHourBehavior(),
 };
 
 export const useLivechatEnterprise = () => {
-	const businessHourType = useSetting('Livechat_business_hour_type') as string;
+	const businessHourType = useSetting<'Single' | 'Multiple'>('Livechat_business_hour_type', 'Single');
 	const { data: hasLicense = false } = useHasLicenseModule('livechat-enterprise');
 
 	useEffect(() => {
 		if (businessHourType && hasLicense) {
-			businessHourManager.registerBusinessHourBehavior(businessHours[businessHourType.toLowerCase()]);
+			businessHourManager.registerBusinessHourBehavior(businessHours[businessHourType.toLowerCase() as 'single' | 'multiple']);
 		}
 	}, [businessHourType, hasLicense]);
 };
